Move CASTBuilder to c_simple_ast_bulider
